// src/server/member/types.ts
import type { Position } from '#/server/shared/types';

export type Streak = { count: number; type: 'win' | 'lose' };

export type Member = {
  id: string;
  name: string;
  mainPosition: Position;
  subPositions: Position[];
  mmr: number;
  isTemporary: boolean;
  createdAt: string;
  streak: Streak | null;
};

export type MemberInput = {
  name: string;
  mainPosition: Position;
  subPositions: Position[];
  isTemporary: boolean;
};